Fixed digital input acquisition
NI DAQmx refuses to just give you a binary array. You have to get integers where each line is in a bit location corresponding to the line number. Even if not all the lines are being acquired; lines that are not acquired are just represented by zeros in the bits of the integer.